Added .htaccess file
The .htaccess file has a URL Rewrite Engine. When a URL request is sent to the server from the browser, the .htaccess file scans the URL for a particular string pattern. If the string pattern matches, the .htaccess file 'rewrites' the URL in a particular way. The server then processes this URL request as normal and returns the output to the browser.

In this .htaccess file, if the browser submits 'localhost/index.php' (for example) to the server the .htaccess file will rerite it as 'localhost/server/view/index.php' which is where all our 'View' files are held.